requirements.txt
anyio==3.6.2
async-timeout==4.0.2
boto3==1.26.27
botocore==1.29.27
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
dnspython==2.2.1
fastapi==0.88.0
h11==0.14.0
httptools==0.5.0
idna==3.4
jmespath==1.0.1
psutil==5.9.4
pydantic==1.10.2
pymongo==4.3.3
python-dateutil==2.8.2
python-dotenv==0.21.1
PyYAML==6.0
redis==4.5.1
requests==2.28.1
rq==1.12.0
s3transfer==0.6.0
six==1.16.0
sniffio==1.3.0
starlette==0.22.0
typing_extensions==4.4.0
urllib3==1.26.13
uvicorn==0.20.0
uvloop==0.17.0
watchfiles==0.18.1
websockets==10.4
